improve test runner (#30)
